+++ monocle/github/graphql.py
@@ -20,7 +20,7 @@ import requests
 from time import sleep
 from datetime import datetime
 
-from tenacity import retry, stop_after_attempt, wait_fixed
+from tenacity import retry, wait_exponential, before_log
 
 
 class GithubGraphQLQuery(object):
@@ -85,7 +85,9 @@ class GithubGraphQLQuery(object):
         data = self.query(qdata, skip_get_rate_limit=True)
         return data['data']['rateLimit']
 
-    @retry(stop=stop_after_attempt(6), wait=wait_fixed(10))
+    @retry(
+        before=before_log(log, logging.INFO), wait=wait_exponential(min=10, max=300),
+    )
     def query(self, qdata, skip_get_rate_limit=False, ignore_not_found=False):
         if not skip_get_rate_limit:
             if self.query_count % self.get_rate_limit_rate == 0: