+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
+- [crawler] github use exponential tenacity retry
 - [crawler] follow [GitHub abuse rate limits guidelines](https://developer.github.com/v3/guides/best-practices-for-integrators/#dealing-with-abuse-rate-limits):
   obey to the Retry-After header for the GitHub api calls, sleep 1s
   between calls and set the User-Agent header to change-metrics/monocle.